wip: refactor rom module
