+++ crates/mes-core/src/features/json.rs
@@ -1,4 +1,4 @@
-use crate::{cartridge::Header, error::Error};
+use crate::{error::Error, rom::ines::Header};
 
 pub fn serialize_rom_header(bytes: &[u8]) -> Result<String, Error> {
     let header = Header::try_from_bytes(bytes)?;
