+++ crates/mes-core/src/lib.rs
@@ -1,11 +1,11 @@
 pub mod apu;
 pub mod bus;
-pub mod cartridge;
 pub mod controller;
 pub mod cpu;
 pub mod error;
 pub mod mappers;
 pub mod ppu;
+pub mod rom;
 pub mod utils;
 
 mod features;
